+++ src/phpcs-report/worker.ts
@@ -234,6 +234,8 @@ export class Worker {
 
         // Send the document to be handled.
         if (phpcsProcess.stdin.writable) {
+            // Write the input file path before the content so PHPCS can utilize it.
+            phpcsProcess.stdin.write('phpcs_input_file: ' + request.documentPath + '\n');
             phpcsProcess.stdin.end(request.documentContent);
         }